chore: divide keras_model into keras_model_with_save_compile and keras_model
